core: Change SUIT command handlers scope
Change the scope of operation for all SUIT command handlers, so they are
called once for each active component.
This change is required for further implementation of complex conditions
and directives, such as is-dependency or process-dependency.
Additionally, it creates a better separation between modules and moves
the component list handling into the suit_seq_exec module.

There are two exceptions from this change:
 - suit_directive_set_current_components: this operation is not oriented
   on active components, but changes the list of them.
 - suit_directive_override_parameters: this operation may change the
   execution state (soft-failure parameter) instead of
   component-specific parameters, so the loop over parameters and
   components must stay inside the command implementation.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>