Dry run only executed for sequences to run in a given context
This commit changes which sequences are run during a dry run
(until now all sequences were run).

The dry run now only runs the sequence currently being processed
and these which are executed after the sequence currently processed.

Additionally:
- The dry run is not run when processing the PARSE sequence
- The sequences run from the bootloader context - the INSTALL sequence
  and any sequences that come after it are NOT executed when processing
  the sequences executed in application context (DEPENDENCY_RESOLUTION
  and PAYLOAD_FETCH).

The reason for this change is that when performing a dry run in
application context the dry run for sequences such as INSTALL
could fail, although the manifest is valid (due to different
supported configurations).
The same issue can occur the other way round.

Ref: NCSDK-25996

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>