cddl: expand constrains
The SUIT processor must be able to fulfill the following scenario:
Components:
1. APP_LOCAL_1
2. APP_LOCAL_2
3. APP_LOCAL_3
4. RAD_LOCAL_1
5. RAD_LOCAL_2
6. NORDIC_TOP
7. CAND_MFST
8. CACHE_POOL
9. dev-cfg [0]
10. dev-cfg [1]
11. dev-var

The following constrains were modified:
- integrated payload limit: 5 -> 6
- component limit: 5 -> 12
- dependency limit: 4 -> 7

The number of SUIT compoments was increased to 12, which means that it
can fit all 11 required components with one slot of margin remaining.

The number of integrated payloads was increased to 6, which is enough to
fit all manifest integrated payloads.

The number of dependencies was increased to 7, which is enough to
process all 6 dependency manifests plus APP_recovery.

Signed-off-by: Rafał Kuźnia <rafal.kuznia@nordicsemi.no>